+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/presenters/JsonFormFragmentPresenter.java
@@ -134,7 +134,7 @@ public class JsonFormFragmentPresenter extends MvpBasePresenter<JsonFormFragment
         } else if (childAt instanceof CustomTextView) {
             CustomTextView customTextView = (CustomTextView) childAt;
             String type = (String) childAt.getTag(R.id.type);
-            if (!TextUtils.isEmpty(type) && type.equals(JsonFormConstants.NUMBER_SELECTOR)) {
+            if (!TextUtils.isEmpty(type) && type.equals(JsonFormConstants.NUMBERS_SELECTOR)) {
                 ValidationStatus validationStatus = NumberSelectorFactory.validate(formFragmentView, customTextView);
                 if (!validationStatus.isValid()) {
                     if (requestFocus) validationStatus.requestAttention();
@@ -342,7 +342,7 @@ public class JsonFormFragmentPresenter extends MvpBasePresenter<JsonFormFragment
                     showInformationDialog(v);
                 }
                 break;
-            case JsonFormConstants.NUMBER_SELECTOR:
+            case JsonFormConstants.NUMBERS_SELECTOR:
                 createNumberSelector(v);
                 break;
             case JsonFormConstants.SPINNER:
@@ -522,10 +522,10 @@ public class JsonFormFragmentPresenter extends MvpBasePresenter<JsonFormFragment
 
                     if (isChecked) {
                         if (exclusiveSet.contains(childKey)) {
-                            getView().unCheckAllExcept(parentKey, childKey);
+                            getView().unCheckAllExcept(parentKey, childKey, compoundButton);
                         } else {
                             for (String excludeKey : exclusiveSet) {
-                                getView().unCheck(parentKey, excludeKey);
+                                getView().unCheck(parentKey, excludeKey, compoundButton);
                             }
                         }
                     }
@@ -548,7 +548,7 @@ public class JsonFormFragmentPresenter extends MvpBasePresenter<JsonFormFragment
                 popup = false;
             }
 
-            getView().unCheckAllExcept(parentKey, childKey);
+            getView().unCheckAllExcept(parentKey, childKey, compoundButton);
 
             getView().writeValue(mStepName, parentKey, childKey, openMrsEntityParent,
                     openMrsEntity, openMrsEntityId, popup);
@@ -572,7 +572,7 @@ public class JsonFormFragmentPresenter extends MvpBasePresenter<JsonFormFragment
                     openMrsEntityId, popup);
         }
 
-        if (JsonFormConstants.NUMBER_SELECTOR.equals(type)) {
+        if (JsonFormConstants.NUMBERS_SELECTOR.equals(type)) {
             NumberSelectorFactory.setBackgrounds(customTextView);
             NumberSelectorFactory.setSelectedTextViews(customTextView);
             NumberSelectorFactory.setSelectedTextViewText((String) parent.getItemAtPosition(position));