+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/utils/Utils.java
@@ -3,8 +3,22 @@ package com.vijay.jsonwizard.utils;
 import android.content.Context;
 import android.widget.Toast;
 
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 public class Utils {
     public static void showToast(Context context, String message) {
         Toast.makeText(context, message, Toast.LENGTH_LONG).show();
+    }
+
+    public static Date getDateFromString(String dtStart) {
+        SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
+        try {
+            Date date = format.parse(dtStart);
+            return date;
+        } catch (ParseException e) {
+            return null;
+        }
     }
 }